Use a different nav layer for Linux and Mac hosts
On macOS, GUI + arrow keys are used to achieve page up, page down, home
and end. For a consistent experience on Linux, this behaviour is
achieved through the use of mod morphs.

Switching between hosts is now done with a macro that effectively
changes the base layer, so that the appropriate nav layer can be
accessed.