Workaround to cancel sticky keys by tapping `&mo`
Well, kinda... this reconfigures my thumb `&mo` keys into a hold-tap:

  * Hold: Activate the layer (just like `&mo`)
  * Tap:  Send F24

When tapped, any active sticky keys are applied to the F24 keypress and
thus "cancelled". (Although the F24 keypress is sent to the host, it is
unlikely to be bound to anything in the applications I use.)

Unfortunately, this functionality clears sticky keys only when tapping
the key, it does not clear them when activating the layer (via hold).

Related ZMK GitHub thread:
https://github.com/zmkfirmware/zmk/issues/834

This is an attempt to have one-shot functionality closer to Callum's:
https://github.com/callum-oakley/qmk_firmware/blob/master/users/callum/readme.md#oneshot-modifiers